Refactor and cleanup auth code
