+++ backends/parking/src/config/config.ts
@@ -1,4 +1,10 @@
-import { Configuration } from 'src/config/Configuration'
+import { AuthModuleConfig, buildAuthModuleConfig } from '@parker/nest-utils'
+
+export interface Configuration {
+  environment: 'dev' | 'prod'
+  port: number
+  auth: AuthModuleConfig
+}
 
 const port = Number(process.env['PORT'] ?? 3501)
 const hostName: string = process.env['HOST_NAME'] ?? 'http://localhost'
@@ -6,9 +12,5 @@ const hostName: string = process.env['HOST_NAME'] ?? 'http://localhost'
 export const config: Configuration = {
   environment: 'dev',
   port,
-  auth: {
-    apiUrl: `${hostName}:${port}`,
-    supertokensUrl: process.env['SUPERTOKENS_CORE_URL'],
-    websiteDomain: process.env['PARKER_WEB_URL'],
-  },
+  auth: buildAuthModuleConfig({ apiUrl: `${hostName}:${port}` }),
 }
