+++ backends/parking/src/config/index.ts
@@ -1,2 +1,1 @@
 export * from './config'
-export * from './Configuration'
