+++ backends/parking/src/test/buildTestApp.ts
@@ -1,6 +1,6 @@
 import { INestApplication } from '@nestjs/common'
 import { Test } from '@nestjs/testing'
-import { MicroserviceAuthModule, NestAppBuilder, SuperTokensExceptionFilter } from '@parker/nest-utils'
+import { MicroserviceAuthModule, NestAppBuilder } from '@parker/nest-utils'
 import { config } from 'src/config'
 import { ParkingSpotModule } from 'src/domain/parkingSpot'
 import { ParkingSpotBookingModule } from 'src/domain/parkingSpotBooking'
@@ -10,8 +10,7 @@ export const buildTestApp = async (): Promise<INestApplication> => {
     imports: [ParkingSpotModule, ParkingSpotBookingModule, MicroserviceAuthModule.forRoot(config.auth)],
   }).compile()
   const app = moduleRef.createNestApplication()
-  NestAppBuilder.configureApp(app)
-  app.useGlobalFilters(new SuperTokensExceptionFilter())
+  NestAppBuilder.configureApp(app, undefined)
   await app.init()
   return app
 }
