+++ backends/places/src/main.ts
@@ -4,8 +4,7 @@ import * as dotenv from 'dotenv'
 dotenv.config()
 
 import { Module } from '@nestjs/common'
-import { MicroserviceAuthModule, NestAppBuilder, NestAppRunner, SuperTokensExceptionFilter } from '@parker/nest-utils'
-import supertokens from 'supertokens-node'
+import { MicroserviceAuthModule, NestAppBuilder, NestAppRunner } from '@parker/nest-utils'
 import { config } from 'src/config'
 import { PlaceSuggestionsModule } from 'src/domain/placeSuggestions'
 import { Logger } from '@parker/logging'
@@ -16,14 +15,7 @@ import { Logger } from '@parker/logging'
 class AppModule {}
 
 const bootstrap = async (port: number): Promise<void> => {
-  const app = await NestAppBuilder.build(AppModule)
-  // TODO: probably this CORS stuff is only necessary once this api starts being called by browsers?
-  app.enableCors({
-    origin: ['http://localhost:3000'], // TODO: URL of the website domain
-    allowedHeaders: ['content-type', ...supertokens.getAllCORSHeaders()],
-    credentials: true,
-  })
-  app.useGlobalFilters(new SuperTokensExceptionFilter())
+  const app = await NestAppBuilder.build(AppModule, config.auth.websiteDomain)
   await NestAppRunner.run(app, port)
 }
