+++ backends/places/src/test/buildTestApp.ts
@@ -1,6 +1,6 @@
 import { INestApplication } from '@nestjs/common'
 import { Test } from '@nestjs/testing'
-import { MicroserviceAuthModule, NestAppBuilder, SuperTokensExceptionFilter } from '@parker/nest-utils'
+import { MicroserviceAuthModule, NestAppBuilder } from '@parker/nest-utils'
 import { config } from 'src/config'
 import { PlaceSuggestionsModule } from 'src/domain/placeSuggestions'
 
@@ -9,8 +9,7 @@ export const buildTestApp = async (): Promise<INestApplication> => {
     imports: [PlaceSuggestionsModule, MicroserviceAuthModule.forRoot(config.auth)],
   }).compile()
   const app = moduleRef.createNestApplication()
-  NestAppBuilder.configureApp(app)
-  app.useGlobalFilters(new SuperTokensExceptionFilter())
+  NestAppBuilder.configureApp(app, undefined)
   await app.init()
   return app
 }
