--- backends/user/src/config/Configuration.ts
@@ -1,7 +1,0 @@
-import { AuthModuleConfig } from '@parker/nest-utils'
-
-export interface Configuration {
-  environment: 'dev' | 'prod'
-  port: number
-  auth: AuthModuleConfig
-}
