+++ backends/user/src/config/config.ts
@@ -1,14 +1,16 @@
-import { Configuration } from 'src/config/Configuration'
+import { AuthModuleConfig, buildAuthModuleConfig } from '@parker/nest-utils'
 
 const port = Number(process.env['PORT'] ?? 3503)
 const hostName: string = process.env['HOST_NAME'] ?? 'http://localhost'
 
+export interface Configuration {
+  environment: 'dev' | 'prod'
+  port: number
+  auth: AuthModuleConfig
+}
+
 export const config: Configuration = {
   environment: 'dev',
   port,
-  auth: {
-    apiUrl: `${hostName}:${port}`,
-    supertokensUrl: process.env['SUPERTOKENS_CORE_URL'],
-    websiteDomain: process.env['PARKER_WEB_URL'],
-  },
+  auth: buildAuthModuleConfig({ apiUrl: `${hostName}:${port}` }),
 }
