+++ backends/user/src/main.ts
@@ -4,8 +4,7 @@ import * as dotenv from 'dotenv'
 dotenv.config()
 
 import { Module } from '@nestjs/common'
-import { LoginServiceAuthModule, NestAppBuilder, NestAppRunner, SuperTokensExceptionFilter } from '@parker/nest-utils'
-import supertokens from 'supertokens-node'
+import { LoginServiceAuthModule, NestAppBuilder, NestAppRunner } from '@parker/nest-utils'
 import { config } from 'src/config'
 import { Logger } from '@parker/logging'
 
@@ -15,14 +14,7 @@ import { Logger } from '@parker/logging'
 class AppModule {}
 
 const bootstrap = async (port: number): Promise<void> => {
-  const app = await NestAppBuilder.build(AppModule)
-  // TODO: probably this CORS stuff is only necessary once this api starts being called by browsers?
-  app.enableCors({
-    origin: ['http://localhost:3000'], // TODO: URL of the website domain
-    allowedHeaders: ['content-type', ...supertokens.getAllCORSHeaders()],
-    credentials: true,
-  })
-  app.useGlobalFilters(new SuperTokensExceptionFilter())
+  const app = await NestAppBuilder.build(AppModule, config.auth.websiteDomain)
   await NestAppRunner.run(app, port)
 }
