+++ backends/user/src/test/buildTestApp.ts
@@ -1,6 +1,6 @@
 import { INestApplication } from '@nestjs/common'
 import { Test } from '@nestjs/testing'
-import { LoginServiceAuthModule, NestAppBuilder, SuperTokensExceptionFilter } from '@parker/nest-utils'
+import { LoginServiceAuthModule, NestAppBuilder } from '@parker/nest-utils'
 import { config } from 'src/config'
 
 export const buildTestApp = async (): Promise<INestApplication> => {
@@ -8,8 +8,7 @@ export const buildTestApp = async (): Promise<INestApplication> => {
     imports: [LoginServiceAuthModule.forRoot(config.auth)],
   }).compile()
   const app = moduleRef.createNestApplication()
-  NestAppBuilder.configureApp(app)
-  app.useGlobalFilters(new SuperTokensExceptionFilter())
+  NestAppBuilder.configureApp(app, undefined)
   await app.init()
   return app
 }
