+++ packages/nest-utils/src/index.ts
@@ -1,6 +1,10 @@
+export * from './lib/auth/AuthGuard'
+export * from './lib/auth/AuthModuleConfig'
+export * from './lib/auth/LoginServiceAuthModule'
+export * from './lib/auth/MicroserviceAuthModule'
+export * from './lib/auth/Session'
 export * from './lib/BaseController'
 export * from './lib/HttpStatus'
 export * from './lib/NestAppBuilder'
 export * from './lib/NestAppRunner'
 export * from './lib/Endpoint'
-export * from './lib/auth'
