+++ packages/nest-utils/src/lib/auth/LoginServiceAuthModule.ts
@@ -25,9 +25,9 @@ export class LoginServiceAuthModule implements NestModule {
       appInfo: {
         appName: 'Parker',
         apiDomain: apiUrl,
-        websiteDomain: websiteDomain ?? 'http://localhost:3000',
+        websiteDomain,
       },
-      connectionURI: supertokensUrl ?? 'http://localhost:3567',
+      connectionURI: supertokensUrl,
       apiKey,
       recipeList: [EmailPassword.init(), Session.init()],
     }
