--- packages/nest-utils/src/lib/auth/index.ts
@@ -1,10 +1,0 @@
-export * from './AuthConfig'
-export * from './AuthConfigInjectionToken'
-export * from './AuthGuard'
-export * from './AuthEndpoints'
-export * from './AuthModuleConfig'
-export * from './LoginServiceAuthModule'
-export * from './MicroserviceAuthModule'
-export * from './Session'
-export * from './SuperTokensExceptionFilter'
-export * from './SuperTokensService'
